feat: move SemesterSelector to components directory and update imports
